# perf(cbor): for encoding numbers, bigints, and dates