This pull request offers performance improvements to the encoding of numbers, BigInts and dates into the cbor format. The current version of cbor appears to beat msgpack in small numbers and then do worse in larger numbers. These changes causes cbor to do slightly worse than itself in smaller numbers but overall better than msgpack with all ranges of numbers.

I did try similar changes with Uint8Array and Arrays, but Arrays did worse overall. For Uint8Arrays, I was able to get a performance boost for small lengths of data, but very easily worse for large lengths of data.

## Benchmark Source Code
- `std/cbor/deno.json`
```json
{
  "name": "@std/cbo",
  "version": "0.1.2",
  "exports": {
    ".": "./mod.ts",
    "./array-encoder-stream": "./array_encoder_stream.ts",
    "./byte-encoder-stream": "./byte_encoder_stream.ts",
    "./decode-cbor-sequence": "./decode_cbor_sequence.ts",
    "./decode-cbor": "./decode_cbor.ts",
    "./encode-cbor-sequence": "./encode_cbor_sequence.ts",
    "./encode-cbor": "./encode_cbor.ts",
    "./map-encoder-stream": "./map_encoder_stream.ts",
    "./sequence-decoder-stream": "./sequence_decoder_stream.ts",
    "./sequence-encoder-stream": "./sequence_encoder_stream.ts",
    "./tag": "./tag.ts",
    "./text-encoder-stream": "./text_encoder_stream.ts",
    "./types": "./types.ts"
  }
}

```
- `std/bench`
```ts
import { encodeCbor as encodeCborOld } from "@std/cbor";
import { encodeCbor as encodeCborNew } from "@std/cbo";
import { encode as encodeMsg } from "@std/msgpack";

function random(start: number, end: number): number {
  return Math.floor(Math.random() * (end - start) + start);
}

const Byte0 = random(0, 24);
const Byte1 = random(24, 2 ** 8);
const Byte2 = random(2 ** 8, 2 ** 16);
const Byte4 = random(2 ** 16, 2 ** 32);
const Byte8 = random(2 ** 32, 2 ** 64);
const Now = new Date();

Deno.bench(
  { name: "Cbor Old: Number - 0 Byte Length", group: "N:Byte0" },
  () => {
    encodeCborOld(Byte0);
  },
);

Deno.bench(
  { name: "Cbor New: Number - 0 Byte Length", group: "N:Byte0" },
  () => {
    encodeCborNew(Byte0);
  },
);

Deno.bench(
  { name: "MsgPack: Number - 0 Byte Length", group: "N:Byte0" },
  () => {
    encodeMsg(Byte0);
  },
);

Deno.bench(
  { name: "Cbor Old: Number - 1 Byte Length", group: "N:Byte1" },
  () => {
    encodeCborOld(Byte1);
  },
);

Deno.bench(
  { name: "Cbor New: Number - 1 Byte Length", group: "N:Byte1" },
  () => {
    encodeCborNew(Byte1);
  },
);

Deno.bench(
  { name: "MsgPack: Number - 1 Byte Length", group: "N:Byte1" },
  () => {
    encodeMsg(Byte1);
  },
);

Deno.bench(
  { name: "Cbor Old: Number - 2 Byte Length", group: "N:Byte2" },
  () => {
    encodeCborOld(Byte2);
  },
);

Deno.bench(
  { name: "Cbor New: Number - 2 Byte Length", group: "N:Byte2" },
  () => {
    encodeCborNew(Byte2);
  },
);

Deno.bench(
  { name: "MsgPack: Number - 2 Byte Length", group: "N:Byte2" },
  () => {
    encodeMsg(Byte2);
  },
);

Deno.bench(
  { name: "Cbor Old: Number - 4 Byte Length", group: "N:Byte4" },
  () => {
    encodeCborOld(Byte4);
  },
);

Deno.bench(
  { name: "Cbor New: Number - 4 Byte Length", group: "N:Byte4" },
  () => {
    encodeCborNew(Byte4);
  },
);

Deno.bench(
  { name: "MsgPack: Number - 4 Byte Length", group: "N:Byte4" },
  () => {
    encodeMsg(Byte4);
  },
);

Deno.bench(
  { name: "Cbor Old: Number - 8 Byte Length", group: "N:Byte8" },
  () => {
    encodeCborOld(Byte8);
  },
);

Deno.bench(
  { name: "Cbor New: Number - 8 Byte Length", group: "N:Byte8" },
  () => {
    encodeCborNew(Byte8);
  },
);

Deno.bench(
  { name: "MsgPack: Number - 8 Byte Length", group: "N:Byte8" },
  () => {
    encodeMsg(Byte8);
  },
);

Deno.bench(
  { name: "Cbor Old: BigInt - 0 Byte Length", group: "B:Byte0" },
  () => {
    encodeCborOld(BigInt(Byte0));
  },
);

Deno.bench(
  { name: "Cbor New: BigInt - 0 Byte Length", group: "B:Byte0" },
  () => {
    encodeCborNew(BigInt(Byte0));
  },
);

Deno.bench(
  { name: "MsgPack: BigInt - 0 Byte Length", group: "B:Byte0" },
  () => {
    encodeMsg(BigInt(Byte0));
  },
);

Deno.bench(
  { name: "Cbor Old: BigInt - 1 Byte Length", group: "B:Byte1" },
  () => {
    encodeCborOld(BigInt(Byte1));
  },
);

Deno.bench(
  { name: "Cbor New: BigInt - 1 Byte Length", group: "B:Byte1" },
  () => {
    encodeCborNew(BigInt(Byte1));
  },
);

Deno.bench(
  { name: "MsgPack: BigInt - 1 Byte Length", group: "B:Byte1" },
  () => {
    encodeMsg(BigInt(Byte1));
  },
);

Deno.bench(
  { name: "Cbor Old: BigInt - 2 Byte Length", group: "B:Byte2" },
  () => {
    encodeCborOld(BigInt(Byte2));
  },
);

Deno.bench(
  { name: "Cbor New: BigInt - 2 Byte Length", group: "B:Byte2" },
  () => {
    encodeCborNew(BigInt(Byte2));
  },
);

Deno.bench(
  { name: "MsgPack: BigInt - 2 Byte Length", group: "B:Byte2" },
  () => {
    encodeMsg(BigInt(Byte2));
  },
);

Deno.bench(
  { name: "Cbor Old: BigInt - 4 Byte Length", group: "B:Byte4" },
  () => {
    encodeCborOld(BigInt(Byte4));
  },
);

Deno.bench(
  { name: "Cbor New: BigInt - 4 Byte Length", group: "B:Byte4" },
  () => {
    encodeCborNew(BigInt(Byte4));
  },
);

Deno.bench(
  { name: "MsgPack: BigInt - 4 Byte Length", group: "B:Byte4" },
  () => {
    encodeMsg(BigInt(Byte4));
  },
);

Deno.bench(
  { name: "Cbor Old: BigInt - 8 Byte Length", group: "B:Byte8" },
  () => {
    encodeCborOld(BigInt(Byte8));
  },
);

Deno.bench(
  { name: "Cbor New: BigInt - 8 Byte Length", group: "B:Byte8" },
  () => {
    encodeCborNew(BigInt(Byte8));
  },
);

Deno.bench(
  { name: "MsgPack: BigInt - 8 Byte Length", group: "B:Byte8" },
  () => {
    encodeMsg(BigInt(Byte8));
  },
);

Deno.bench({ name: "Cbor Old: Date", group: "Date" }, () => {
  encodeCborOld(Now);
});

Deno.bench({ name: "Cbor New: Date", group: "Date" }, () => {
  encodeCborNew(Now);
});
```

## Benchmark Results
```
Check file:///Users/soul/Projects/std/bench
    CPU | Apple M1
Runtime | Deno 2.1.1 (aarch64-apple-darwin)

file:///Users/soul/Projects/std/bench

benchmark                          time/iter (avg)        iter/s      (min … max)           p75      p99     p995
---------------------------------- ----------------------------- --------------------- --------------------------

group N:Byte0
Cbor Old: Number - 0 Byte Length          160.5 ns     6,229,000 (145.1 ns … 182.8 ns) 166.6 ns 176.6 ns 182.2 ns
Cbor New: Number - 0 Byte Length          207.3 ns     4,824,000 (198.0 ns … 248.8 ns) 208.8 ns 225.6 ns 229.5 ns
MsgPack: Number - 0 Byte Length           320.6 ns     3,119,000 (299.5 ns … 376.7 ns) 326.7 ns 351.6 ns 376.7 ns

summary
  Cbor Old: Number - 0 Byte Length
     1.29x faster than Cbor New: Number - 0 Byte Length
     2.00x faster than MsgPack: Number - 0 Byte Length

group N:Byte1
Cbor Old: Number - 1 Byte Length          160.9 ns     6,217,000 (146.7 ns … 178.7 ns) 166.3 ns 174.2 ns 176.3 ns
Cbor New: Number - 1 Byte Length          205.9 ns     4,856,000 (197.2 ns … 220.2 ns) 207.6 ns 219.1 ns 219.7 ns
MsgPack: Number - 1 Byte Length           320.2 ns     3,123,000 (302.0 ns … 341.3 ns) 325.8 ns 339.4 ns 341.3 ns

summary
  Cbor Old: Number - 1 Byte Length
     1.28x faster than Cbor New: Number - 1 Byte Length
     1.99x faster than MsgPack: Number - 1 Byte Length

group N:Byte2
Cbor Old: Number - 2 Byte Length          717.7 ns     1,393,000 (699.8 ns … 739.6 ns) 726.8 ns 739.6 ns 739.6 ns
Cbor New: Number - 2 Byte Length          206.7 ns     4,838,000 (195.8 ns … 219.9 ns) 208.6 ns 217.3 ns 219.0 ns
MsgPack: Number - 2 Byte Length           322.6 ns     3,100,000 (303.6 ns … 350.5 ns) 329.9 ns 344.9 ns 350.5 ns

summary
  Cbor New: Number - 2 Byte Length
     1.56x faster than MsgPack: Number - 2 Byte Length
     3.47x faster than Cbor Old: Number - 2 Byte Length

group N:Byte4
Cbor Old: Number - 4 Byte Length          748.9 ns     1,335,000 (719.2 ns … 815.2 ns) 757.4 ns 815.2 ns 815.2 ns
Cbor New: Number - 4 Byte Length          206.2 ns     4,849,000 (194.9 ns … 227.7 ns) 208.0 ns 222.8 ns 223.8 ns
MsgPack: Number - 4 Byte Length           325.7 ns     3,071,000 (303.1 ns … 352.2 ns) 334.0 ns 352.2 ns 352.2 ns

summary
  Cbor New: Number - 4 Byte Length
     1.58x faster than MsgPack: Number - 4 Byte Length
     3.63x faster than Cbor Old: Number - 4 Byte Length

group N:Byte8
Cbor Old: Number - 8 Byte Length          922.3 ns     1,084,000 (768.3 ns …   2.4 µs) 876.0 ns   2.4 µs   2.4 µs
Cbor New: Number - 8 Byte Length          198.5 ns     5,039,000 (180.2 ns … 226.7 ns) 204.1 ns 221.8 ns 223.5 ns
MsgPack: Number - 8 Byte Length           331.5 ns     3,017,000 (303.4 ns … 700.6 ns) 334.4 ns 458.1 ns 700.6 ns

summary
  Cbor New: Number - 8 Byte Length
     1.67x faster than MsgPack: Number - 8 Byte Length
     4.65x faster than Cbor Old: Number - 8 Byte Length

group B:Byte0
Cbor Old: BigInt - 0 Byte Length          288.7 ns     3,464,000 (182.5 ns …   1.9 µs) 308.8 ns 993.1 ns   1.9 µs
Cbor New: BigInt - 0 Byte Length          275.0 ns     3,637,000 (237.8 ns … 586.0 ns) 280.5 ns 507.0 ns 586.0 ns
MsgPack: BigInt - 0 Byte Length           343.9 ns     2,908,000 (313.6 ns … 727.5 ns) 349.0 ns 454.8 ns 727.5 ns

summary
  Cbor New: BigInt - 0 Byte Length
     1.05x faster than Cbor Old: BigInt - 0 Byte Length
     1.25x faster than MsgPack: BigInt - 0 Byte Length

group B:Byte1
Cbor Old: BigInt - 1 Byte Length          196.4 ns     5,093,000 (186.7 ns … 216.7 ns) 198.4 ns 211.8 ns 212.7 ns
Cbor New: BigInt - 1 Byte Length          248.2 ns     4,029,000 (244.1 ns … 269.2 ns) 248.9 ns 259.0 ns 259.5 ns
MsgPack: BigInt - 1 Byte Length           330.6 ns     3,025,000 (310.0 ns … 375.1 ns) 336.6 ns 373.8 ns 375.1 ns

summary
  Cbor Old: BigInt - 1 Byte Length
     1.26x faster than Cbor New: BigInt - 1 Byte Length
     1.68x faster than MsgPack: BigInt - 1 Byte Length

group B:Byte2
Cbor Old: BigInt - 2 Byte Length          781.9 ns     1,279,000 (765.0 ns … 810.7 ns) 792.0 ns 810.7 ns 810.7 ns
Cbor New: BigInt - 2 Byte Length          251.6 ns     3,975,000 (245.8 ns … 331.7 ns) 251.5 ns 276.9 ns 321.1 ns
MsgPack: BigInt - 2 Byte Length           333.9 ns     2,995,000 (311.0 ns … 454.0 ns) 337.7 ns 439.9 ns 454.0 ns

summary
  Cbor New: BigInt - 2 Byte Length
     1.33x faster than MsgPack: BigInt - 2 Byte Length
     3.11x faster than Cbor Old: BigInt - 2 Byte Length

group B:Byte4
Cbor Old: BigInt - 4 Byte Length          790.2 ns     1,266,000 (763.0 ns … 832.6 ns) 799.9 ns 832.6 ns 832.6 ns
Cbor New: BigInt - 4 Byte Length          256.7 ns     3,896,000 (252.1 ns … 271.9 ns) 257.6 ns 268.9 ns 270.6 ns
MsgPack: BigInt - 4 Byte Length           330.4 ns     3,027,000 (311.5 ns … 382.4 ns) 336.6 ns 370.6 ns 382.4 ns

summary
  Cbor New: BigInt - 4 Byte Length
     1.29x faster than MsgPack: BigInt - 4 Byte Length
     3.08x faster than Cbor Old: BigInt - 4 Byte Length

group B:Byte8
Cbor Old: BigInt - 8 Byte Length          795.9 ns     1,256,000 (760.0 ns … 881.6 ns) 797.6 ns 881.6 ns 881.6 ns
Cbor New: BigInt - 8 Byte Length          250.0 ns     4,000,000 (228.1 ns … 399.2 ns) 254.5 ns 372.2 ns 372.3 ns
MsgPack: BigInt - 8 Byte Length           388.6 ns     2,573,000 (327.6 ns … 750.9 ns) 395.3 ns 702.9 ns 750.9 ns

summary
  Cbor New: BigInt - 8 Byte Length
     1.55x faster than MsgPack: BigInt - 8 Byte Length
     3.18x faster than Cbor Old: BigInt - 8 Byte Length

group Date
Cbor Old: Date                              1.1 µs       887,200 (  1.1 µs …   1.3 µs)   1.1 µs   1.3 µs   1.3 µs
Cbor New: Date                            401.8 ns     2,489,000 (386.6 ns … 470.9 ns) 407.7 ns 447.2 ns 470.9 ns

summary
  Cbor New: Date
     2.81x faster than Cbor Old: Date
```